requirements.txt
python-dotenv
llama-index-core==0.10.26
llama-index-llms-openai==0.1.19
llama-index-llms-groq==0.1.3
llama-index-llms-huggingface==0.1.4
llama-index-llms-ollama==0.1.3
llama-index-llms-gemini==0.1.7

llama-index-embeddings-huggingface==0.1.3
llama-index-embeddings-ollama==0.1.2

llama-index-tools-duckduckgo==0.1.0
llama-index-vector-stores-chroma==0.1.5
vllm==0.4.0
fastapi
easydict==1.9.0
feedparser==6.0.11
schedule==1.2.1
sendgrid
markdown
pyvis==0.3.2
chainlit==1.1.101
pypdf
sentence_transformers==2.7.0
scholarly==1.7.11